+++ src/App.tsx
@@ -1,7 +1,7 @@
 
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
-import { withErrorBoundary } from './ignitus-Shared/Components/ErrorBoundary';
+import { withErrorBoundary } from './ignitus-Shared/Components/errorBoundary';
 import { ProtectedRoutes } from './ignitus-Routes/protectedRoutes';
 // eslint-disable-next-line import/named
 import { PublicRoutes } from './ignitus-Routes/publicRoutes';
