+++ src/ignitus-Common-Auth-Student-Prof/Components/commonLoginUI.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { withErrorBoundary } from '../../ignitus-Shared/Components/ErrorBoundary';
+import { withErrorBoundary } from '../../ignitus-Shared/Components/errorBoundary';
 import * as t from './Constants';
 import '../Styles/style.scss';
