+++ src/ignitus-Partners/Components/Partners.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable import/no-unresolved */
 import React from 'react';
-import { withErrorBoundary } from '../../ignitus-Shared/Components/ErrorBoundary';
+import { withErrorBoundary } from '../../ignitus-Shared/Components/errorBoundary';
 import { Data } from '../constants';
 import { PureContributors } from '../../ignitus-Contributors/Components/Contributors';
