+++ src/ignitus-Student-Dashboard/Components/Dashboard.js
@@ -4,7 +4,7 @@
 /* eslint-disable react/jsx-filename-extension */
 import React from 'react';
 import '../Styles/style.scss';
-import { withErrorBoundary } from '../../ignitus-Shared/Components/ErrorBoundary';
+import { withErrorBoundary } from '../../ignitus-Shared/Components/errorBoundary';
 
 const Dashboard = () => {
   let email = '';
