+++ src/ignitus-StudentLogin/Components/Login.js
@@ -4,7 +4,7 @@
 import React, { useState } from 'react';
 import isEmpty from 'lodash/isEmpty';
 import PropTypes from 'prop-types';
-import { withErrorBoundary } from '../../ignitus-Shared/Components/ErrorBoundary';
+import { withErrorBoundary } from '../../ignitus-Shared/Components/errorBoundary';
 import SharedLogin from '../../ignitus-Shared/Components/Login/Login';
 import { statePayload } from '../../ignitus-Shared/Components/Login/Constants';
